backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
groq
python-dotenv
pymongo[srv]
motor
